+++ tests/copyright/add_notice/test_5.m
@@ -1,0 +1,2 @@
+%This is a corner case, here we have no space between
+%the % and the start of the comment